feat: parallel batch [T20260509-5]
Tasks:
- T20260509-5: Modularize crates/orbit-cli/src/command/mcp/setup.rs (1,964 lines)
