+++ crates/orbit-cli/src/command/mcp/setup/format/mod.rs
@@ -1,0 +1,5 @@
+mod json;
+mod toml;
+
+pub(super) use self::json::*;
+pub(super) use self::toml::*;
