+++ crates/orbit-cli/src/command/mcp/setup/mod.rs
@@ -1,0 +1,12 @@
+mod args;
+mod dispatch;
+mod format;
+mod providers;
+#[cfg(test)]
+mod test_support;
+mod workspace;
+
+#[allow(unused_imports)]
+pub use args::ScopeArg;
+pub(crate) use args::init_auto_for_workspace;
+pub use args::{InitArgs, RemoveArgs};
