+++ crates/orbit-cli/src/command/mcp/setup/providers/common.rs
@@ -1,0 +1,3 @@
+pub(super) fn server_args() -> Vec<String> {
+    vec!["mcp".to_string(), "serve".to_string()]
+}
