+++ crates/orbit-cli/src/command/mcp/setup/providers/mod.rs
@@ -1,0 +1,10 @@
+mod claude;
+mod codex;
+mod common;
+mod gemini;
+mod simple_json;
+
+pub(super) use self::claude::{apply_claude_init, apply_claude_remove};
+pub(super) use self::codex::{apply_codex_init, apply_codex_remove};
+pub(super) use self::gemini::{apply_gemini_init, apply_gemini_remove};
+pub(super) use self::simple_json::{apply_simple_json_init, apply_simple_json_remove};
