+++ crates/orbit-cli/src/command/mcp/setup/test_support.rs
@@ -1,0 +1,32 @@
+use std::ffi::OsString;
+use std::sync::Mutex;
+
+pub(super) static ENV_LOCK: Mutex<()> = Mutex::new(());
+
+pub(super) struct EnvVarGuard {
+    key: &'static str,
+    previous: Option<OsString>,
+}
+
+impl EnvVarGuard {
+    pub(super) fn set(key: &'static str, value: OsString) -> Self {
+        let previous = std::env::var_os(key);
+        unsafe {
+            std::env::set_var(key, value);
+        }
+        Self { key, previous }
+    }
+}
+
+impl Drop for EnvVarGuard {
+    fn drop(&mut self) {
+        match &self.previous {
+            Some(value) => unsafe {
+                std::env::set_var(self.key, value);
+            },
+            None => unsafe {
+                std::env::remove_var(self.key);
+            },
+        }
+    }
+}
